'#1866: Externalize BookmarkAndKey class.
